+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.6.1](https://github.com/BBVAEngineering/ember-cli-workbox/compare/v0.6.0...v0.6.1) (2020-01-14)
+
+
+### Bug Fixes
+
+* **sw:** do not create empty SW when disabled ([af1cf2d](https://github.com/BBVAEngineering/ember-cli-workbox/commit/af1cf2d)), closes [#78](https://github.com/BBVAEngineering/ember-cli-workbox/issues/78)
+
+# [0.6.0](https://github.com/BBVAEngineering/ember-cli-workbox/compare/v0.5.2...v0.6.0) (2019-12-16)
+
+
+### Features
+
+* add importScriptsTransform option ([440f1f1](https://github.com/BBVAEngineering/ember-cli-workbox/commit/440f1f1))
+
 ## [0.5.2](https://github.com/BBVAEngineering/ember-cli-workbox/compare/v0.5.1...v0.5.2) (2019-03-08)